Refactor folder structure reducers
