+++ front-end/src/reducers/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
-import restReducer from './restReducer';
-import { clientReducer, connectedClientReducer } from './clientReducer';
-import subReducer from './subReducer';
+import restReducer from './API/restReducer';
+import { clientReducer, connectedClientReducer } from './API/clientReducer';
+import subReducer from './API/subReducer';
 import joinFormReducer from './joinFormReducer';
 import readyReducer from './readyReducer';
